requirements.txt
autopep8==1.4.4
bleach==3.1.0
Django==2.2.5
django-crispy-forms==1.7.2
django-markdownify==0.8.0
Markdown==3.1.1
pycodestyle==2.5.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
webencodings==0.5.1
celery
celery[redis]
django-celery-results
redis
factory_boy